+++ src/function/matrix/map.js
@@ -1,5 +1,5 @@
-import { applyCallback } from '../../utils/applyCallback.js'
-import { arraySize, broadcastSizes, broadcastTo, get } from '../../utils/array.js'
+import { optimizeCallback } from '../../utils/optimizeCallback.js'
+import { arraySize, broadcastSizes, broadcastTo, get, recurse } from '../../utils/array.js'
 import { factory } from '../../utils/factory.js'
 
 const name = 'map'
@@ -143,36 +143,14 @@ export const createMap = /* #__PURE__ */ factory(name, dependencies, ({ typed })
       return 0
     }
   }
-})
-
-/**
+  /**
  * Map for a multi dimensional array
  * @param {Array} array
  * @param {Function} callback
  * @return {Array}
  * @private
  */
-function _mapArray (array, callback) {
-  return _recurse(array, [], array, callback)
-}
-
-/**
- * Recursive function to map a multi-dimensional array.
- *
- * @param {*} value - The current value being processed in the array.
- * @param {Array} index - The index of the current value being processed in the array.
- * @param {Array} array - The array being processed.
- * @param {Function} callback - Function that produces the element of the new Array, taking three arguments: the value of the element, the index of the element, and the Array being processed.
- * @returns {*} The new array with each element being the result of the callback function.
- */
-function _recurse (value, index, array, callback) {
-  if (Array.isArray(value)) {
-    return value.map(function (child, i) {
-      // we create a copy of the index array and append the new index value
-      return _recurse(child, index.concat(i), array, callback)
-    })
-  } else {
-    // invoke the callback function with the right number of arguments
-    return applyCallback(callback, value, index, array, 'map')
+  function _mapArray (array, callback) {
+    return recurse(array, [], array, optimizeCallback(callback, array, name))
   }
-}
+})